restructure oauth redux cycle
