+++ src/modules/user/actions.js
@@ -1,0 +1,12 @@
+import * as UserAPI from './utils.js';
+
+export const RECEIVE_USER = 'RECEIVE_USER';
+
+export const receiveUser = user => ({
+  type: RECEIVE_USER,
+  user
+});
+
+export const loginUser = () => dispatch => {
+  UserAPI.authenticateUser(dispatch);
+};
