+++ src/modules/user/reducer.js
@@ -1,0 +1,16 @@
+import { merge } from 'lodash';
+import { RECEIVE_USER } from "./actions.js";
+
+let _defaultState = {};
+
+const userReducer = (state = _defaultState, action) => {
+  Object.freeze(state);
+  switch(action.type) {
+    case RECEIVE_USER:
+      return action.user;
+    default:
+      return state;
+  }
+};
+
+export default userReducer;
