+++ src3/reducers/root_reducer.js
@@ -6,8 +6,8 @@ import recommendedReducer from './recommended_reducer';
 import searchResultReducer from './search_result_reducer';
 import subscriptionsReducer from './subscriptions_reducer';
 import trendingReducer from './trending_reducer';
-import userReducer from './user_reducer';
-import settingReducer from './setting_reducer';
+// import userReducer from './user_reducer';
+// import settingReducer from './setting_reducer';
 import playlistsReducer from './playlists_reducer';
 import playerDetailsReducer from './player_details_reducer';
 
@@ -19,8 +19,8 @@ const rootReducer = combineReducers({
   searchResult: searchResultReducer,
   subscriptions: subscriptionsReducer,
   trending: trendingReducer,
-  user: userReducer,
-  setting: settingReducer,
+  // user: userReducer,
+  // setting: settingReducer,
   playlists: playlistsReducer,
   playerDetails: playerDetailsReducer
 });
